refactor(auth): replace Flags with AccessLevel and rename Authorization → AuthContext
* Rename **Authorization.{hpp,cpp}** to **AuthContext.{hpp,cpp}**; update all
  includes and CMake targets accordingly.
* Introduce `AuthContext::AccessLevel` (`Unprotected`, `User`, `AdminUser`)
  and drop the bit-flag-based `Flags`, simplifying permission checks.
* Re-implement `AuthContext::isAuthorized()` with clearer logic, bot detection,
  and more descriptive `Result::Reason` values.
* Update every call-site to the new API
  ( *TgBotApiImpl, ModuleManagement, OnInlineQuery, Restart, SpamBlockManager* ).
* Rename `CommandModule::isEnforced()` → `isPrivileged()` and adjust module
  loading / command registration to honour the new access levels.
* Improve diagnostics:
* Better log messages in `DynCommandModule`.
* `LuaCommandModule` now validates `command.name`/`description` and logs a
  helpful error when they’re missing.

+191  −274 net lines. No functional behaviour changes outside of the
  authorization refactor.

+++ src/include/SharedMalloc.hpp
@@ -11,7 +11,7 @@
 
 struct SharedMalloc {
     using offset_type = int;
-    using size_type = size_t;
+    using size_type = long;
     using data_type = std::uint8_t;
 
     struct Parent {
